pyproject.toml
[project]
name = "repl-nix-teacherdashboardpro"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "django>=5.1.4",
    "django-cors-headers>=4.6.0",
    "djangorestframework>=3.15.2",
    "mysqlclient>=2.2.6",
    "python-dotenv>=1.0.1",
    "psycopg2-binary>=2.9.10",
]
